+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 
+## v2.3.0
+### 2024-08-26
+
+The VarInfoBase.get_missing_variable_attributes method has been added to allow
+someone to get metadata attributes from the configuration file for variables
+that are absent from a file. An example usage is when a CF Convention grid
+mapping variable is missing from a source file.
+The VarInfoBase.get_references_for_attribute method has been added to retrieve
+all unique variable references contained in a single metadata attribute for a
+list of variables. For example, retrieving all references listed under the
+coordinates metadata attribute.
+
 ## v2.2.2
 ### 2024-07-16